--- Miso/Sources/Application/Flow/Flows/DemoStep.swift
@@ -1,9 +1,0 @@
-import Foundation
-import RxFlow
-
-
-enum DemoStep: Step {
-    
-    case loginIsRequired
-    case mainIsRequired
-}
