--- Miso/Sources/Application/Flow/Flows/LoginStepper.swift
@@ -1,17 +1,0 @@
-import Foundation
-import RxFlow
-import RxCocoa
-import RxSwift
-
-class LoginStepper: Stepper {
-    
-    let steps = PublishRelay<Step>()
-    private let disposeBag = DisposeBag()
-    
-    init() {
-    }
-    
-    var initialStep: Step {
-        return DemoStep.loginIsRequired
-    }
-}
