--- Miso/Sources/Application/Flow/Flows/MainStepper.swift
@@ -1,18 +1,0 @@
-import Foundation
-import RxFlow
-import RxCocoa
-import RxSwift
-
-class MainStepper: Stepper {
-    
-    let steps = PublishRelay<Step>()
-    
-    private let disposeBag = DisposeBag()
-    
-    init() {
-    }
-    
-    var initialStep: Step {
-        return DemoStep.mainIsRequired
-    }
-}
