+++ Miso/Sources/Presentation/Auth/Signup/ViewModel/SignupVM.swift
@@ -1,0 +1,8 @@
+//
+//  SignupVM.swift
+//  Miso
+//
+//  Created by 박준서 on 11/19/23.
+//
+
+import Foundation
