+++ Miso/Sources/Presentation/Main/ViewModel/MainVM.swift
@@ -1,0 +1,8 @@
+//
+//  MainVM.swift
+//  Miso
+//
+//  Created by 박준서 on 11/19/23.
+//
+
+import Foundation
