+++ Miso/Sources/Presentation/Search/ViewModel/SearchVM.swift
@@ -1,0 +1,8 @@
+//
+//  SearchVM.swift
+//  Miso
+//
+//  Created by 박준서 on 11/19/23.
+//
+
+import Foundation
